docs(collapsible-section): improve documentations and examples
