+++ src/components/collapsible-section/examples/collapsible-section-actions.tsx
@@ -2,7 +2,16 @@ import { Action } from '@limetech/lime-elements';
 import { Component, h, State } from '@stencil/core';
 
 /**
- * Example with actions
+ * Header actions
+ * An array of actions can be given to the component.
+ * The actions will be rendered as icon buttons inside the Header.
+ * They can allow for additional functionality like edit, delete, etc.
+ *
+ * :::important
+ * The actions should be about the entire section or group of content,
+ * and affect it as a whole. This is due to how they are displayed along
+ * with the header of the group.
+ * :::
  */
 @Component({
     tag: 'limel-example-collapsible-section-actions',
@@ -13,20 +22,24 @@ export class CollapsibleSectionActionsExample {
     private action: Action = {
         id: '0',
         icon: 'test_tube',
+        label: 'No action clicked',
     };
 
     private actions = [
         {
             id: '1',
+            label: 'Delete',
             icon: 'trash',
         },
         {
             id: '2',
+            label: 'Update',
             icon: 'available_updates',
             disabled: true,
         },
         {
             id: '3',
+            label: 'More',
             icon: 'menu_2',
         },
     ];
@@ -38,7 +51,19 @@ export class CollapsibleSectionActionsExample {
                 actions={this.actions}
                 onAction={this.handleAction}
             >
-                <limel-icon name={this.action.icon} size="large" />
+                <p>Last clicked action:</p>
+                <ul>
+                    <li>
+                        ID: <code>{this.action.id}</code>
+                    </li>
+                    <li>
+                        Label: <code>{this.action.label}</code>
+                    </li>
+                    <li>
+                        Icon: <code>{this.action.icon}</code>
+                        <limel-icon name={this.action.icon} size="x-small" />
+                    </li>
+                </ul>
             </limel-collapsible-section>
         );
     }
